% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{can_continue}
\alias{can_continue}
\title{Check if story can continue}
\usage{
can_continue(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Check if story can continue
}
\seealso{
Other API: 
\code{\link{choose_choice_index}()},
\code{\link{continue_maximally}()},
\code{\link{continue_normally}()}
}
\concept{API}
