% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{choose_choice_index}
\alias{choose_choice_index}
\title{Make a choice}
\usage{
choose_choice_index(js, idx)
}
\arguments{
\item{js}{javascript context to load story into}

\item{idx}{integer index of choice}
}
\description{
Make a choice
}
\seealso{
Other API: 
\code{\link{can_continue}()},
\code{\link{continue_maximally}()},
\code{\link{continue_normally}()}
}
\concept{API}
