% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{continue_maximally}
\alias{continue_maximally}
\title{Continue story until a choice is hit}
\usage{
continue_maximally(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Continue story until a choice is hit
}
\seealso{
Other API: 
\code{\link{can_continue}()},
\code{\link{choose_choice_index}()},
\code{\link{continue_normally}()}
}
\concept{API}
