% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{continue_normally}
\alias{continue_normally}
\title{Continue story one step at a time}
\usage{
continue_normally(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Continue story one step at a time
}
\seealso{
Other API: 
\code{\link{can_continue}()},
\code{\link{choose_choice_index}()},
\code{\link{continue_maximally}()}
}
\concept{API}
