% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{current_choices_length}
\alias{current_choices_length}
\title{Get number of current choices}
\usage{
current_choices_length(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Get number of current choices
}
\keyword{internal}
