% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{current_choices_text}
\alias{current_choices_text}
\title{Get text of current choices}
\usage{
current_choices_text(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Get text of current choices
}
\keyword{internal}
