% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inkar}
\alias{inkar}
\title{INKAR data}
\format{
A list containing connections to the tables in the local INKAR database.
\describe{
\item{\verb{_indikatoren}}{ID (official INKAR ID), name, domain and
description of individual indicators.}
\item{\verb{_regionen}}{ID (internal ID in the local database),
spatial reference, official code and name of individual regions.}
\item{other tables, e.g. \code{absolutzahlen}}{Values by region and
time. Each table contains values for a certain INKAR domain (Bereich).
\code{region_id} corresponds to the \code{id} column in the \verb{_regionen} table.
Column names correspond to the \code{id} column in the \verb{_indikatoren} table.}
}
The DBI connection to the local database is stored as an attribute of the
\code{inkar} object and can be accessed with \code{attr(inkar, "con")}.
}
\usage{
inkar
}
\description{
Tables in the INKAR database
}
\keyword{datasets}
