% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inkar-db.R
\name{inkar_db_build}
\alias{inkar_db_build}
\title{Import INKAR data into a local DuckDB database}
\usage{
inkar_db_build(overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{overwrite}{Overwrite existing database}

\item{verbose}{Should SQL statements for building the local database be printed?}
}
\description{
Import INKAR data into a local DuckDB database
}
