% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kreise}
\alias{kreise}
\title{NUTS-3 regions (districts/\emph{Kreise})}
\format{
An object of class \code{data.frame} with 401 rows and 3 columns.
}
\usage{
kreise
}
\description{
Lookup table with NUTS-3 code and \emph{Kreiskennziffer} for all 401 NUTS-3 regions in Germany
}
\keyword{datasets}
