% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.r
\name{load_story}
\alias{load_story}
\title{Load Ink story from JSON}
\usage{
load_story(js, filepath)
}
\arguments{
\item{js}{javascript context to load story into}

\item{filepath}{path to JSON file}
}
\description{
Load Ink story from JSON
}
