% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.r
\name{remove_context}
\alias{remove_context}
\title{Remove a javascript context}
\usage{
remove_context(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Remove a javascript context
}
