% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{rename_with_inkar_indicators}
\alias{rename_with_inkar_indicators}
\title{Rename columns to INKAR indicator names instead of x123 etc.}
\usage{
rename_with_inkar_indicators(.data)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}
}
\value{
An object of the same type as \code{.data}. The output has the following
properties:
\itemize{
\item Rows are not affected.
\item Column names are changed; column order is preserved.
\item Data frame attributes are preserved.
\item Groups are updated to reflect new names.
}
}
\description{
Rename columns to INKAR indicator names instead of x123 etc.
}
\examples{
\dontrun{
library(dplyr)
inkar$bevoelkerung |>
  filter(zeitbezug == 2019) |>
  select(region_id, x111:x115) |>
  rename_with_inkar_indicators()
}
}
