% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{version_ink}
\alias{version_ink}
\title{Version of Ink engine}
\usage{
version_ink()
}
\description{
Version of Ink engine
}
\seealso{
Other version: 
\code{\link{version_storystate}()},
\code{\link{version_story}()}
}
\concept{version}
