% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{version_story}
\alias{version_story}
\title{Version of Ink story}
\usage{
version_story(filepath)
}
\arguments{
\item{filepath}{path to JSON file}
}
\description{
Version of Ink story
}
\seealso{
Other version: 
\code{\link{version_ink}()},
\code{\link{version_storystate}()}
}
\concept{version}
