% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.r
\name{version_storystate}
\alias{version_storystate}
\title{Version of Ink story state}
\usage{
version_storystate(js)
}
\arguments{
\item{js}{javascript context to load story into}
}
\description{
Version of Ink story state
}
\seealso{
Other version: 
\code{\link{version_ink}()},
\code{\link{version_story}()}
}
\concept{version}
\keyword{internal}
